IoT_Mini_Project_Group_H/requirements.txt
paho-mqtt
python-dotenv
requests
supabase
